Initial Commit
Cute little dungeon crawler on the system console. Try to reach the exit typing in movement keys while avoiding monsters and traps. Uses system("cls") to clear the console and make it look prettier. I know it is looked down upon by I do not intend to use it past this tier of programming.